Reorganize frontend files
