--- frontend/src/LoginPage.js
@@ -1,15 +1,0 @@
-import React from "react";
-import LoginBody from "./components/LoginBody";
-import Footer from "./components/Footer";
-import Navbar from "./components/Navbar";
-
-const LoginPage = () => {
-  return (
-    <>
-      <LoginBody />
-      <Footer />
-    </>
-  );
-};
-
-export default LoginPage;
