+++ frontend/src/components/Footer.js
@@ -1,5 +1,5 @@
 import React from "react";
-import '../footer.css';
+import '../css/footer.css';
 
 const Footer = () => {
   return (
