+++ frontend/src/pages/LoginPage.js
@@ -1,0 +1,14 @@
+import React from "react";
+import LoginBody from "../components/LoginBody";
+import Footer from "../components/Footer";
+import Navbar from "../components/Navbar";
+
+const LoginPage = () => {
+  return (
+    <>
+      <LoginBody />
+    </>
+  );
+};
+
+export default LoginPage;
