--- frontend/src/signup.css
@@ -1,44 +1,0 @@
-.container {
-    display: "flex";
-    flex-direction: "column",
-    alignItems: "center",
-    justifyContent: "center",
-    minHeight: "100vh",
-    padding: "1rem",
-  }
-.title {
-    fontSize: "1.8rem",
-    fontWeight: "bold",
-    marginBottom: "1rem",
-  }
-.form {
-    width: "100%",
-    maxWidth: "20rem",
-    display: "flex",
-    flexDirection: "column",
-    gap: "1rem",
-  }
-.input {
-    width: "100%",
-    padding: "0.5rem",
-    border: "1px solid #ccc",
-    borderRadius: "0.5rem",
-  }
-.button {
-    width: "100%",
-    backgroundColor: "#007bff",
-    color: "white",
-    padding: "0.75rem",
-    border: "none",
-    borderRadius: "0.5rem",
-    cursor: "pointer",
-  }
-.link {
-    marginTop: "1rem",
-    fontSize: "0.9rem",
-  }
-.loginLink {
-    color: "#007bff",
-    cursor: "pointer",
-    textDecoration: "underline",
-  }
